Move some blockdevmgr methods into devices module
Move initialize and wipe_blockdevs. initialize is getting bigger and
needs room to breathe. It depends on wipe_blockdevs. All methods in this
module are about blockdevs before they are owned by Stratis.
